move merge functions
