+++ projects/ngrx-clean-forms/src/lib/utils.ts
@@ -8,6 +8,9 @@ import {
     FormGroupControlSummaries,
     FormGroupControlUpdates,
     Validator,
+    FormGroupErrors,
+    FormArrayErrors,
+    FormControlErrors,
 } from './types';
 
 export function mapFormControlStates<TControls extends FormControls, R>(
@@ -65,4 +68,69 @@ export function validatorOf<T>(fn: ValidatorFn): Validator<T> {
             enabled: !control.disabled,
             disabled: control.disabled,
         } as AbstractControl);
+}
+
+export function mergeFormControlErrors(...errors: FormControlErrors[]): FormControlErrors | null {
+    return errors.reduce((e1, e2) => {
+        if (!e1 && !e2) {
+            return null;
+        }
+
+        if (!e1) {
+            return e2;
+        }
+
+        if (!e2) {
+            return e1;
+        }
+
+        return {
+            ...e1,
+            ...e2,
+        };
+    }, null);
+}
+
+export function mergeFormGroupErrors<TControls extends FormControls>(
+    ...errors: FormGroupErrors<TControls>[]
+): FormGroupErrors<TControls> | null {
+    return errors.reduce((group1, group2) => {
+        if (!group1 && !group2) {
+            return null;
+        }
+
+        if (!group1) {
+            return group2;
+        }
+
+        if (!group2) {
+            return group1;
+        }
+
+        return {
+            ...group1,
+            ...group2,
+            ...Object.keys(group1)
+                .filter(key1 => Object.keys(group2).find(key2 => key1 === key2))
+                .map(key => ({
+                    [key]: mergeFormControlErrors(group1[key], group2[key]),
+                }))
+                .reduce(
+                    (e1, e2) => ({
+                        ...e1,
+                        ...e2,
+                    }),
+                    {}
+                ),
+        };
+    }, null);
+}
+
+export function mergeFormArrayErrors(...errors: FormArrayErrors[]): FormArrayErrors | null {
+    const mergedErrors = errors
+        .filter(Boolean)
+        .reduce((arr1, arr2) => (arr1.length >= arr2.length ? arr1 : arr2), [])
+        .map((_, i) => mergeFormControlErrors(...errors.map(arrayErrors => arrayErrors[i])));
+
+    return mergedErrors.filter(Boolean).length > 0 ? mergedErrors : null;
 }
